Improvements to Scaling
Added a couple new events so you can find out when channels are created
or destroyed for a platform.  Also adjusted the autoscaling to be a
little bit more conservative.

When channels are destroyed they now process the rest of the queue
before being destroyed, this eliminates the case where a teardown may
have happened to a channel with queued notifications, where those
notifications may have been forever lost!