Big changes on Apple notification management
There were several changes to the underlying base Channel and Service
classes to, but many changes to the Apple Channel, for how it manages
notifications, and how it decides when all queued messages have been
dealt with and it can stop.  This should really improve the durability
of the library, especially when calling .Stop() or stopping a service
immediately after queueing some notifications!

Added some Logging statements in the service that may be useful to some.

Added nuget version 1.1.0.0

Fixes #69, Fixes #75