Multiple event handler execution WP
When a Windows Phone registration is created, a ServiceRegistration
object is instantiated for each type of notification (there are 6)
however because the PushService instance is the same, event handlers are
assigned multiple times.  This causes an event handler to be executed
multiple times for the same notification.

To be able to address the issues I used a C# pattern that doesn't allow
the same event handler to be assigned to an event.