fixed duplicate registration error message
This is a quick fix to prevent confusing users, but there is a larger problem that should be fixed: If a client registers and then goes down (for example because the registration response does not arrive due to a network error), when the client retries registration, by default, it fails because the previous registration hasn't timed out. This default can be overridden by setting raiseErrorOnDuplicateRegistrations=false. However, this default seems backwards. Additionally, having to deviate from the default means you can't use extension methods like RegisterGcmService.