Fix bug GCM Authorization Failed #574
'Authorization' Header was being appended to for each notification in
the service queue. This caused multiple notifications to fail as the
header got appended for each call. New code is simpler and should be
better performing. The auth headers are set only once at the creation of
the service connection