add APNS and GCM proxy support (#1)
* Update PushShart using a Proxy when connecting to Google Cloud Messaging (GCM)

* Options for Proxy Usage

Adds properties and methods to use an HTTP-proxy in your connection, either with default network credentials or specified credentials.

* Connection via proxy

A method to tunnel the TCP-connection via a pre-configured HTTP-Proxy has been added. If Configuration.UseProxy is set to true, a CONNECT-request is sent using the proxy adress specified. The socket is then extracted from the stream of this request's response (using the System.Reflection-Library) and used to create a TcpClient-object. All traffic sent through this TcpClient-object will pass through the proxy used to send the initial CONNECT-Request.

NOTE:
The method 'getClientViaHTTPProxy' uses code that was not written by the author of this commit. The original code can be found here:
https://web.archive.org/web/20160317134733/https://nitormobiledevelopment.wordpress.com/2013/08/13/push-sharp-using-proxy/

* Update ApnsConnection.cs

* Update ApnsConfiguration.cs

* apns wit proxy withou socket hijack

* Handling proxy connection errors on APNS

* Using System.Net missing in GcmConfiguration

* Http Proxy with Basic authentication without WebRequest in APNS

* System.Net was missing in GcmConfiguration

* Using http proxy with basic auth (not using socket hijacking) in APNS

* New class ProxyHelper to permit proxy connect in any place using a TcpClient

* Using ProxyHelper in ApnsConnection

* Supporting proxy in ApnsFeedbackService.

* Changes in ApnsFeedbackService to return the tokens list intead of raising events

* add proxy usage section to readme